nix_bot/requirements.txt
python-telegram-bot
sqlalchemy
solana
cryptography
